refactor: extract bed form builders, move delete confirmation to BedsView, align error handling
- Extract shared buildBedFieldInput into bedFormInput.ts (was duplicated in createBedMutation.ts)
- Drop CREATE_BED_MUTATION / UPDATE_BED_MUTATION aliases; use generated constants directly
- Move ConfirmationModal + delete confirmation state from BedTable to BedsView
- Add errorPolicy: 'all' + OperationInfo check to handleDeleteBeds (matches handleClone pattern)
- Keep BedTable as a pure presentation component with a simple onDeleteBeds callback
